Version 2 fromdanceguy(gmail) <danceguy@gmail.com> toJ2 <j2sentient@gmail.com> date9 July 2010 12:05 subjectRe: bowling.jar
James:

I was up late last night playing with Junit and bowling program.

Attached is a revised version of the code, that works better, and passes
more test cases.
This jar file has "V2" in the name to mean "version 2", but it extracts
into a
./Bowling  folder (PS: before we left lab last night, I
refactored/renamed the project name
to be Bowling )

I also added code to throw an exception if the length of the pinsDown[]
array
has the wrong length.

I had trouble initially getting the code to compile, because the
.classpath file from lab
was referring to a JDK 1.6, but my machine currently has only 1.5 JDK
installed.
The way I fixed this problem was to create a new Bowling2 project,
and copy the source files into Bowling2 Project, delete old Bowling
project, and
rename Bowling2 to Bowling project (renamed in ECLIPSE by refactoring).

I'm having trouble with ECLIPSE 3.1 (an old version) use of
ANNOTATATIONS.
(@Test, @Before, @After , etc in front of methods.).  I tried to fix
this problem by creating a
lib folder, and adding the newest Junit.4.8.2.jar file to my classpath,
BUT
it did NOT completely fix the problem.  One of the test cases, I wanted
to use annotations
to look for an expected IndexOutOfBoundsException, but I could not, so
I had to do a more clumsy way.  You will see my attempt at using
annotations
in the attached file.

I put a few comments (BUG FIX) to list the problem areas that I
encoutered.

Last, I also created a new libsrc\ folder, and put the JAR file that
contains the
source code for JUNIT there.  I then used "ATTACH SOURCE" feature in
ECLIPSE
to attach the junit.4.8.2.src.jar file, which is helpful for debugging.

I also added a few more test cases to the project.

--- later...
-- dennis bednar